+++ server.R
@@ -63,9 +63,35 @@ function(input, output, session) {
       map_dfr(chronique.ouverture, Type = "Mesures", typemesure = "Piézométrie", skipvalue = 1, nbcolonnes = 3, separateur_colonnes = ";", separateur_decimales = ".", typefichier = ".csv", typedate = "dmy_hms", typecapteur = "Hobo", nomfichier = T)
   })
 
-  output$fichiers_bruts_apres_importation <- render_gt(fichiers_bruts_apres_importation())
-  output$files <- render_gt(fichiers_bruts_apres_importation() %>% distinct(nom_fichier))
+  #### Nettoyage ####
+  fichiers_propres_apres_importation <- reactive({
+    req(input$upload_fichiers)
+    fichiers_bruts_apres_importation() %>%
+      separate(nom_fichier, c("chmes_coderhj", "nom_fichier"), sep = "_", ) %>%  # Ajout du code station à partir du nom de fichier
+      select(chmes_coderhj, contains("chmes_capteur"), chmes_date, chmes_heure, chmes_valeur, chmes_unite, chmes_typemesure) %>% 
+      distinct() # Dédoublonnage
+  })
   
-  #### Données initiales ####
+  #### Évaluation ####
+  fichiers_bruts_apres_importation_contexte <- reactive({
+    req(input$upload_fichiers)
+    fichiers_bruts_apres_importation() %>%
+      chronique.contexte()
+  })
+  
+  fichiers_propres_apres_importation_contexte <- reactive({
+    req(input$upload_fichiers)
+    fichiers_propres_apres_importation() %>%
+      chronique.contexte()
+  })
+  
+  #### Affichage ####
+  output$fichiers_bruts_apres_importation <- render_gt(fichiers_bruts_apres_importation() %>% filter(grepl("gouterot", nom_fichier)) %>% head())
+  output$fichiers_propres_apres_importation <- render_gt(fichiers_propres_apres_importation() %>% filter(grepl("gouterot", chmes_coderhj)) %>% head())
+  output$fichiers_propres_apres_importation_test <- render_gt(fichiers_propres_apres_importation() %>% filter(grepl("porf", chmes_coderhj)) %>% head())
+  output$files <- render_gt(fichiers_bruts_apres_importation() %>% distinct(nom_fichier))
+  output$fichiers_bruts_apres_importation_contexte <- render_gt(fichiers_bruts_apres_importation_contexte())
+  output$fichiers_propres_apres_importation_contexte <- render_gt(fichiers_propres_apres_importation_contexte())
+  
   
 }