+++ NEWS.md
@@ -1,4 +1,11 @@
 
+### 0.0.3 - 2025-02-18
+- Nettoyage du code inutilisé de l'UI
+- Extraction du nom de la station à partir du nom du fichier
+- Dédoublonnage des chroniques
+- Calcul et affichage des contextes du jeu de données initial et dédoublonné
+- Affichage des jeux de données initial et dédoublonné
+
 ### 0.0.2 - 2025-02-18
 - Upload fonctionnel de plusieurs fichiers, sans utilisation possible des boutons de réglages
 - Affichage de tableaux de contrôle des noms de fichiers + données à proprement parler (rétrocontrôle)